ai-agent-service/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
duckdb==1.3.0
fastapi==0.116.1
Flask==3.1.0
Flask-Caching==2.3.0
frozenlist==1.5.0
googleapis-common-protos==1.70.0
griffe==1.7.3
grpcio==1.73.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langfuse==2.57.12
markdown2==2.5.2
MarkupSafe==3.0.2
mcp==1.8.1
moviepy==2.1.2
multidict==6.1.0
numpy==2.2.4
openai==1.78.1
openai-agents==0.0.14
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.11
propcache==0.3.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.15.2
slack_sdk==3.34.0
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
tabulate==0.9.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.2
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.4
yarl==1.18.3
zipp==3.23.0
